pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "od-benchmark"
version = "0.1.0"
description = "Object Detection Performance Benchmark with COCO Dataset"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Contributors"}
]
keywords = ["object-detection", "benchmark", "coco", "yolo", "evaluation", "computer-vision"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "ultralytics>=8.0",
    "pycocotools",
    "opencv-python",
    "pillow",
    "pyyaml",
    "pandas",
    "matplotlib",
    "seaborn",
    "tqdm",
    "rich",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "mypy>=1.0",
    "black>=23.0",
    "ruff>=0.1",
]

[project.entry-points."console_scripts"]
od-benchmark = "src.cli:main"


[project.urls]
Homepage = "https://github.com/yourusername/od-benchmark"
Repository = "https://github.com/yourusername/od-benchmark"
Issues = "https://github.com/yourusername/od-benchmark/issues"


[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
exclude = ["tests*", "examples*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml"]


[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311", "py312"]
include = "\\.pyi?$"
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.ruff_cache
  | \.pytest_cache
  | \.temp
  | \.venv
  | build
  | dist
  | outputs
  | models_cache
)/
'''


[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]


[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true


[[tool.mypy.overrides]]
module = [
    "cv2.*",
    "ultralytics.*",
    "torchvision.*",
    "pycocotools.*",
]


ignore_missing_imports = true


[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]

markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]


[tool.coverage.run]
source = ["src"]
omit = [
    "tests/*",
    "*/__init__.py",
]


[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == \"__main__.:",
]
